refactored admin page
